Move XPBD logic into `XpbdSolverPlugin`, add `SolverPlugins`, restructure
