+++ src/dynamics/solver/xpbd/joints/mod.rs
@@ -1,5 +1,7 @@
-mod fixed_angle_constraint;
-mod point_constraint;
+//! XPBD joint constraints.
+
+mod shared;
+pub use shared::{FixedAngleConstraintShared, PointConstraintShared};
 
 mod distance;
 mod fixed;
@@ -7,11 +9,6 @@ mod prismatic;
 mod revolute;
 #[cfg(feature = "3d")]
 mod spherical;
-
-pub mod joint_graph;
-
-pub use fixed_angle_constraint::FixedAngleConstraintShared;
-pub use point_constraint::PointConstraintShared;
 
 pub use distance::DistanceJointSolverData;
 pub use fixed::FixedJointSolverData;
