+++ src/dynamics/solver/xpbd/joints/revolute.rs
@@ -1,8 +1,6 @@
-//! [`RevoluteJoint`] component.
-
+use super::PointConstraintShared;
 use crate::{
     dynamics::solver::{
-        joints::PointConstraintShared,
         solver_body::{SolverBody, SolverBodyInertia},
         xpbd::*,
     },
@@ -10,6 +8,7 @@ use crate::{
 };
 use bevy::prelude::*;
 
+/// Constraint data required by the XPBD constraint solver for a [`RevoluteJoint`].
 #[derive(Component, Clone, Copy, Debug, Default, PartialEq, Reflect)]
 #[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
 #[cfg_attr(feature = "serialize", reflect(Serialize, Deserialize))]
