+++ src/dynamics/solver/xpbd/joints/shared/mod.rs
@@ -1,0 +1,5 @@
+mod fixed_angle_constraint;
+mod point_constraint;
+
+pub use fixed_angle_constraint::FixedAngleConstraintShared;
+pub use point_constraint::PointConstraintShared;
